relocate exception class to its own package
