+++ src/main/kotlin/systems/ajax/englishstudytelegrambot/exception/WordExceptions.kt
@@ -1,3 +1,3 @@
-package systems.ajax.englishstudytelegrambot
+package systems.ajax.englishstudytelegrambot.exception
 
 class WordNotFoundBySpendingExceptions : RuntimeException()
