+++ src/main/kotlin/systems/ajax/englishstudytelegrambot/service/WordService.kt
@@ -2,7 +2,7 @@ package systems.ajax.englishstudytelegrambot.service
 
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
-import systems.ajax.englishstudytelegrambot.WordNotFoundBySpendingExceptions
+import systems.ajax.englishstudytelegrambot.exception.WordNotFoundBySpendingExceptions
 import systems.ajax.englishstudytelegrambot.entity.Word
 import systems.ajax.englishstudytelegrambot.repository.WordRepository
